do not try to upgrade main server if it does not work

+++ library.sh
@@ -40,6 +40,7 @@ upgrade_osrm() {
 	echo 'ok'
 	# copy do live server
 	scp $datadir/$profile/* 10.9.0.1:$datadir/tmp-$profile/
+	if [ $? -ne 0 ]; then exit; fi
 	scp /usr/local/bin/osrm-routed-$profile 10.9.0.1:
 	ssh 10.9.0.1 "rm $datadir/$profile/* && mv $datadir/tmp-$profile/* $datadir/$profile/ && cp -f ~/osrm-routed-$profile /usr/local/bin/ && killall osrm-routed-$profile";
 	oma f epsilon.sk/routing
